+++ NEWS.md
@@ -31,10 +31,16 @@ Here is `calculate_subject_status` a function with custom logic that calculates 
 - Removed the requirement for common_forms to have a 'Name' column. In addition, a 'Name' column can be provided in study data tabs, indicating a common name per row, which will improve the query selector items (#207).
 - Improved branding with new ClinSight Logo, added favicon, and allowed for study logos, when available (#214).
 - Added options to review on form level. With form-level review, subject-level graphics and tiles will be hidden, and all data will be shown in the tables, and review is enable on all rows. All data in a form can be reviewed at once; if a user tries to do so, an additional confirmation will be requested (#198).
+- Added path that updates db's `query_data` table with existing EDC queries upon initialization & update.
 
 ## Bug fixes
 
 - The event label order calculation is now calculated as intended in the rare cases where it needs to be estimated (for example when the order of occurrence of the events differs per patient) (#225).
+
+
+
+
+
 
 # clinsight 0.2.0